+++ Changes.md
@@ -1,6 +1,6 @@
 ## Unreleased
 
-- Shader caching
+- Shader precompiling and caching
 - New preset variables: bass_smooth, mid_smooth, treb_smooth, vol_smooth
 - AMD GPU detection and support for PSVersion=4 (by @OfficialIncubo)
 - Age filter: Only load presets modified within the last X days